Fix lua console, add .lua scripts

+++ Source/Traveller/MemWriteUtils.cpp
@@ -3,7 +3,7 @@
 
 namespace MemWriteUtils
 {
-	int setMemoryPerms(int address, int size, int perms)
+	int setMemoryPerms(uintptr_t address, int size, int perms)
 	{
 		int last = 0;
 		VirtualProtect(reinterpret_cast<void*>(address), static_cast<SIZE_T>(size), static_cast<DWORD>(perms), reinterpret_cast<PDWORD>(&last));
